requirements.txt
dependency-injector
flask
dnspython
pymongo
pyyaml
requests
bcrypt
flask-jwt